Task_5/requirements.txt
# 🚦 Traffic Sign Recognition - Dependencies

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Core ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Application
Flask>=3.0.0
Werkzeug>=3.0.0

# Jupyter notebooks (for development)
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.66.0